feat: running shell out of utils container
This commit moves xudctl shell out of the utils container. So utils will not need to be running all the time. And utils container becomes a utilities bundle which can be used to simplify the bash implementation like parsing toml config file.
