+++ images/utils/Dockerfile
@@ -1,9 +1,5 @@
-FROM python:3-alpine3.10
-WORKDIR /app
-ADD requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-ADD launcher /usr/local/lib/python3.8/site-packages/launcher
-ADD config_parser.py /usr/local/bin/config_parser
-ADD args_parser.py /usr/local/bin/args_parser
-RUN chmod u+x /usr/local/bin/config_parser
-RUN chmod u+x /usr/local/bin/args_parser
+FROM python:3.8-alpine3.10
+RUN pip install toml docker demjson
+ADD . /src
+WORKDIR /src
+RUN python setup.py install
