+++ images/utils/bin/args_parser
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 
 from argparse import ArgumentParser
 import os
@@ -11,6 +11,7 @@ parser.add_argument("--testnet-dir")
 parser.add_argument("--mainnet-dir")
 parser.add_argument("--external-ip")
 parser.add_argument("--backup-dir")
+parser.add_argument("--dev", action="store_true")
 
 try:
     args = parser.parse_args()
