+++ images/utils/bin/run_command
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+print("run your command here!!!")
