+++ images/utils/launcher/banner.txt
