+++ images/utils/launcher/command/base.py
@@ -1,0 +1,14 @@
+import argparse
+
+class ArgumentError(Exception):
+    def __init__(self, message, usage):
+        super().__init__(message)
+        self.usage = usage
+
+
+class ArgumentParser(argparse.ArgumentParser):
+    """
+    https://stackoverflow.com/questions/5943249/python-argparse-and-controlling-overriding-the-exit-status-code
+    """
+    def error(self, message):
+        raise ArgumentError(message, self.format_usage())
