--- images/utils/launcher/shell/__init__.py
@@ -1,1 +1,0 @@
-from .shell import Shell
